Move integrity check into identity module
