+++ feature/identity-api/src/main/java/com/jraska/github/client/identity/IntegrityTrigger.kt
@@ -1,0 +1,5 @@
+package com.jraska.github.client.identity
+
+interface IntegrityTrigger {
+  fun executeIntegrityCheck()
+}
