+++ feature/identity/src/main/java/com/jraska/github/client/identity/IdentityModule.kt
@@ -1,16 +1,18 @@
 package com.jraska.github.client.identity
 
+import com.jraska.github.client.identity.integrity.IntegrityCheck
 import com.jraska.github.client.identity.internal.AddSessionIdInterceptor
 import com.jraska.github.client.identity.internal.AnonymousIdentity
 import com.jraska.github.client.identity.internal.SessionIdProvider
 import com.jraska.github.client.time.TimeProvider
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.IntoSet
 import okhttp3.Interceptor
 import javax.inject.Singleton
 
-@Module
+@Module(includes = [IdentityModule.Declarations::class])
 object IdentityModule {
 
   @Provides
@@ -32,5 +34,11 @@ object IdentityModule {
   @IntoSet
   internal fun addSessionIdInterceptor(identityProvider: IdentityProvider): Interceptor {
     return AddSessionIdInterceptor(identityProvider)
+  }
+
+  @Module
+  interface Declarations {
+    @Binds
+    fun bindIntegrityCheckTrigger(integrityCheck: IntegrityCheck): IntegrityTrigger
   }
 }
