+++ feature/settings/build.gradle
@@ -18,6 +18,7 @@ dependencies {
   implementation project(':core-api')
   implementation project(':core-android-api')
   implementation project(':feature:config-debug-api')
+  implementation project(':feature:identity-api')
   implementation project(':navigation-api')
   implementation project(':feature:ui-common-api')
 
@@ -29,8 +30,6 @@ dependencies {
   implementation 'com.jraska:console-timber-tree:1.2.0'
   implementation 'com.jakewharton.timber:timber:5.0.1'
   implementation okHttp
-
-  implementation 'com.google.android.play:integrity:1.0.2'
 
   implementation dagger
   kapt daggerAnnotationProcessor
