+++ feature/settings/src/main/java/com/jraska/github/client/settings/SettingsViewModel.kt
@@ -6,6 +6,7 @@ import com.jraska.github.client.Owner
 import com.jraska.github.client.analytics.AnalyticsEvent
 import com.jraska.github.client.analytics.EventAnalytics
 import com.jraska.github.client.config.debug.ui.ConfigRowModelProvider
+import com.jraska.github.client.identity.IntegrityTrigger
 import com.jraska.github.client.navigation.Urls
 import javax.inject.Inject
 
@@ -13,7 +14,7 @@ internal class SettingsViewModel @Inject constructor(
   private val eventAnalytics: EventAnalytics,
   private val deepLinkLauncher: DeepLinkLauncher,
   private val rowModelProvider: ConfigRowModelProvider,
-  private val integrityCheck: IntegrityCheck
+  private val integrityTrigger: IntegrityTrigger
 ) : ViewModel() {
   fun onPurchaseSubmitted(value: String) {
     val money = value.toDoubleOrNull() ?: return
@@ -33,7 +34,7 @@ internal class SettingsViewModel @Inject constructor(
   }
 
   fun onIntegrityCheckClicked() {
-    integrityCheck.run()
+    integrityTrigger.executeIntegrityCheck()
   }
 
   companion object {
